bugfix: исправил работу модального окна и показа ингредиента на отдельной странице
