+++ src/components/app/app.module.css
@@ -7,6 +7,9 @@
 .centerContainer {
     width: 1200px;
     margin: auto;
+    display: flex;
+    align-content: center;
+    justify-content: center;
 }
 
 .mainContent {
