+++ src/components/modal-overlay/modal-overlay.module.css
@@ -4,12 +4,6 @@
     height: 100%;
     top: 0;
     left: 0;
-    position: absolute;
-    /*background-color: rgba(255, 255, 255, 0.5);*/
+    position: fixed;
     background-color: rgba(0, 0, 0, 0.6);
-    /*display: flex;*/
-    /*justify-content: center;*/
-    /*align-content: center;*/
-    /*flex-flow:column nowrap;*/
-    /*z-index: 1;*/
 }
